Added option to play system "beep" when errors are reported
I often times forget to check my console to see if there were errors/warnings printed when my files are linted (linting happens in real time as files are edited and changed). So, I added an option to play the system bell whenever errors or warnings are printed for a file.

Simply pass {beep: true} in the options to enable the "beep".

Note that it also adds a dependency on the "beepbeep" module.

This could be extended fairly easily to play a different number of beeps (based on how many errors were found in a file, for exampe), or only play beeps on errors and not warnings, etc.

Closes #36